add simple render page to seperate concerns
