+++ Dockerrun.aws.json
@@ -1,7 +1,7 @@
 {
   "AWSEBDockerrunVersion": "1",
   "Image": {
-    "Name": "handsomedose/fec-picture",
+    "Name": "handsomedose/fec-capstone",
     "Update": "true"
   },
   "Ports": [
