+++ public/index.html
@@ -9,7 +9,7 @@
 </head>
 
 <body>
-  <div id="app">
+  <div id="picture">
   </div>
 
   <script src="bundle.js"></script>
